Extract `NumericInput` to a separate file.
